+++ openciv/engine/exceptions/key_exception.py
@@ -1,0 +1,9 @@
+from openciv.engine.exceptions._base_exception import BaseException
+
+
+class KeyException(BaseException):
+    pass
+
+
+class KeyNotFoundException(KeyException):
+    pass